Update fs dependency to version 2.12.0
Notable changes:

* https://github.com/synrc/fs/pull/32

Silences the following message which appears to Linux users:

"Setting up watches.  Beware: since -r was given, this may take a while!
                                                                       Watches established."

This message although being specific to the inotify backend, due to the
"beware" wording, implies that the user can do something to silence it,
which is not true.