load dependencies in application start callback
When using the new PLUG_EDITOR integration, there was a problem where
not all dependencies were available when the code reloader recompiled
the code.
This lead to missing dependencies when the live reloader joined the
channel.
This change ensures that all we initially load the deps when the
application starts, so that later recompiles don't affect the deps list.